requirements.txt
# Combined Requirements for Daily Briefing Backend Deployment

# Core web framework
fastapi>=0.104.1
uvicorn[standard]>=0.24.0

# Briefing Generator Core Dependencies
requests>=2.31.0
python-dotenv>=1.0.0
google-generativeai>=0.3.0
aiohttp>=3.12.0
schedule>=1.2.0
click>=8.1.0

# Enhanced news parsing
feedparser>=6.0.10

# Web Server Dependencies
jinja2>=3.1.2
python-multipart>=0.0.6
psutil>=5.9.6
aiofiles>=23.2.1
pydantic>=2.5.0
httpx>=0.25.2
